Add a fix to Google Drive for viewing docs.
Especially photos and videos. A white overlay for photos and videos
is still one of the worst possible interface mistakes, and though
Drive does not make it, inverting everything to fix all the other
crap Drive does makes it wrong, so a pinpoint fix for this is the
easier route.