Add Skyscanner settings
Skyscanner is now full of the worst kind of ads : the ones that are
interspersed inside the search results. Thankfully, they are marked
as crap by having a variety of classes as opposed to real results which
don't have one, so we can just remove any element in the result list
that has a class.

Also remove the adbar on the side.